text-extraction-handler/requirements.txt
redis==5.0.1
pypdf==3.17.0
google-genai==1.55.0
